Properly deprecated Timer class
