+++ src/core/time.js
@@ -9,45 +9,4 @@ const now = (typeof window !== 'undefined') && window.performance && window.perf
     performance.now.bind(performance) :
     Date.now;
 
-/**
- * A Timer counts milliseconds from when start() is called until when stop() is called.
- *
- * @ignore
- */
-class Timer {
-    /**
-     * Create a new Timer instance.
-     */
-    constructor() {
-        this._isRunning = false;
-        this._a = 0;
-        this._b = 0;
-    }
-
-    /**
-     * Start the timer.
-     */
-    start() {
-        this._isRunning = true;
-        this._a = now();
-    }
-
-    /**
-     * Stop the timer.
-     */
-    stop() {
-        this._isRunning = false;
-        this._b = now();
-    }
-
-    /**
-     * Get the number of milliseconds that passed between start() and stop() being called.
-     *
-     * @returns {number} The elapsed milliseconds.
-     */
-    getMilliseconds() {
-        return this._b - this._a;
-    }
-}
-
-export { now, Timer };
+export { now };
